+++ test/integration/default/powershell_profile_spec.rb
@@ -14,11 +14,17 @@
 
 # PowerShell profile
 describe file('C:/Users/Vagrant/Documents/WindowsPowerShell/Profile.ps1') do
-  its('content') { should match(/Enable-ChefDKBootstrap/i) }
+  its('content') { should match(/# See chefdk_bootstrap PowerShell Module/) }
+  its('content') { should match(/Enable-ChefDKBootstrap/) }
 
   if ENV['http_proxy']
     its('content') { should match(/Add-Proxy/) }
   else
     its('content') { should_not match(/Add-Proxy/) }
   end
+end
+
+# Make sure ssh is on the path
+describe command('ssh') do
+  it { should exist }
 end