+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # Revision History for chefdk_bootstrap
 
 ## Unreleased
-* Fix #47 - Add kitchen config.yml file to pass through proxy settings (when needed)
+* Fix [#47](https://github.com/Nordstrom/chefdk_bootstrap/issues/47):
+  Add kitchen config.yml file to pass through proxy settings (when needed)
 
 ## 1.8.0
 * Fix [#130](https://github.com/Nordstrom/chefdk_bootstrap/issues/130):