+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 # Revision History for chefdk_bootstrap
 
 ## Unreleased
+* Fix [#134](https://github.com/Nordstrom/chefdk_bootstrap/issues/134):
+  Put SSH on the path so `kitchen login` works
+* Fix [#132](https://github.com/Nordstrom/chefdk_bootstrap/issues/132):
+  Add comments to $Profile.CurrentUserAllHosts PowerShell profile
+* Fix [#63](https://github.com/Nordstrom/chefdk_bootstrap/issues/63):
+  Dry up the cleanup in the Windows bootstrap
+* Fix [#133](https://github.com/Nordstrom/chefdk_bootstrap/issues/133):
+  Clean up berkshelf environment variable from Windows bootstrap
 * Fix [#47](https://github.com/Nordstrom/chefdk_bootstrap/issues/47):
   Add kitchen config.yml file to pass through proxy settings (when needed)